Feature: Added function to seed board with mine
This function places a single mine on the board. If there are no spaces
left then it simply returns.

This approach is better than selecting a random X and Y coordinate
because it allows us to ensure that unless the board is full, running
this function will always increment the number of mines on the board by
1.